anonymous sign on enabled
App signs into firebase with just a userID generated the first time you run the app. If you uninstall and reinstall app, you will get a new token. App also records this userID and a unique phone token that does not change with a reinstall. this token is used for firebase cloud messaging (notifications).
