--- app/src/main/java/com/example/daniel/heymayo/AnonymousAuthActivity.java
@@ -1,111 +1,0 @@
-package com.example.daniel.heymayo;
-
-import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.text.TextUtils;
-import android.util.Log;
-import android.view.View;
-import android.widget.EditText;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.example.daniel.heymayo.models.AnonUser;
-import com.example.daniel.heymayo.models.User;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthCredential;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.EmailAuthProvider;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-
-public class AnonymousAuthActivity extends BaseActivity implements View.OnClickListener {
-
-    private static final String TAG = "AnonymousAuth";
-
-    private FirebaseAuth mAuth;
-    private DatabaseReference mDatabase;
-    private String uid = getUid();
-    public String token = getToken();
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_anonymous_auth);
-
-        mDatabase = FirebaseDatabase.getInstance().getReference();
-        mAuth = FirebaseAuth.getInstance();
-
-        // Click listeners
-        findViewById(R.id.button_anonymous_sign_in).setOnClickListener(this);
-        findViewById(R.id.button_anonymous_sign_out).setOnClickListener(this);
-    }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        FirebaseUser currentUser = mAuth.getCurrentUser();
-        updateUI(currentUser);
-    }
-
-    private void signInAnonymously() {
-        showProgressDialog();
-        mAuth.signInAnonymously().addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            // Sign in success, update UI with the signed-in user's information
-                            Log.d(TAG, "signInAnonymously:success");
-                            FirebaseUser user = mAuth.getCurrentUser();
-                            updateUI(user);
-                        } else {
-                            // If sign in fails, display a message to the user.
-                            Log.w(TAG, "signInAnonymously:failure", task.getException());
-                            Toast.makeText(AnonymousAuthActivity.this, "Authentication failed.", Toast.LENGTH_SHORT).show();
-                            updateUI(null);
-                        }
-                        hideProgressDialog();
-                    }
-                });
-    }
-
-    private void signOut() {
-        mAuth.signOut();
-        updateUI(null);
-    }
-
-    private void updateUI(FirebaseUser user) {
-        hideProgressDialog();
-
-        TextView idView = findViewById(R.id.anonymous_status_id);
-        boolean isSignedIn = (user != null);
-
-        // Status text
-        if (isSignedIn) {
-            idView.setText(getString(R.string.id_fmt, user.getUid()));
-        } else {
-            idView.setText(R.string.signed_out);
-        }
-
-        // Button visibility
-        findViewById(R.id.button_anonymous_sign_in).setEnabled(!isSignedIn);
-        findViewById(R.id.button_anonymous_sign_out).setEnabled(isSignedIn);
-    }
-
-    @Override
-    public void onClick(View v) {
-        int i = v.getId();
-        if (i == R.id.button_anonymous_sign_in) {
-            signInAnonymously();
-        } else if (i == R.id.button_anonymous_sign_out) {
-            signOut();
-        }
-    }
-
-    private void writeNewUser(String userId, String token) {
-        AnonUser user = new AnonUser(uid, token);
-        mDatabase.child("users").child(userId).setValue(user);
-    }
-}
