+++ app/src/main/java/com/example/daniel/heymayo/models/User.java
@@ -5,15 +5,14 @@ import com.google.firebase.database.IgnoreExtraProperties;
 @IgnoreExtraProperties
 public class User {
 
-    public String username;
-    public String email;
     public String tokenID;
+    public String uid;
 
     public User() {}
 
-    public User(String username, String email, String tokenID) {
-        this.username = username;
-        this.email = email;
+    public User(String userId, String tokenID) {
         this.tokenID = tokenID;
+        this.uid = userId;
     }
+
 }
