--- app/src/main/res/layout/activity_anonymous_auth.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.daniel.heymayo.AnonymousAuthActivity">
-
-    <ImageView
-        android:id="@+id/icon"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:contentDescription="@string/desc_firebase_lockup"
-        android:src="@drawable/firebase_lockup_400" />
-
-    <TextView
-        android:id="@+id/title_anonymous"
-        style="@style/TextAppearance.AppCompat.Medium"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/icon"
-        android:layout_marginBottom="8dp"
-        android:text="Anonymous Sign In" />
-
-    <TextView
-        android:id="@+id/anonymous_status_id"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/title_anonymous"
-        android:ellipsize="end"
-        android:maxLines="1"
-        android:text="@string/signed_out" />
-
-
-    <Button
-        android:id="@+id/button_anonymous_sign_in"
-        android:layout_width="160dp"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/anonymous_status_id"
-        android:layout_toEndOf="@+id/field_email"
-        android:text="Sign In" />
-
-    <Button
-        android:id="@+id/button_anonymous_sign_out"
-        android:layout_width="160dp"
-        android:layout_height="wrap_content"
-        android:layout_alignTop="@+id/button_anonymous_sign_in"
-        android:layout_toRightOf="@+id/button_anonymous_sign_in"
-        android:enabled="false"
-        android:text="Sign Out" />
-
-</RelativeLayout>
